W161a: phantom software-double sweep outside mdlFile. Census 26 functions / 49 phantom calls / 19 files -> 14 / 25 / 13; 12 functions fully cleared of phantom helpers, 24 calls removed. Biggest: 0041ABC0 580->444/608 nd426->308 (3 calls), mt_evtEffect 00395E20 1004->860/1072 nd750->636 (5 calls), _h_camp_menu 00155070 1520->1452/1696 nd1156->1053 (3 calls), fclCombineEffect 00418510 732->660/880 nd491->452. Files touched: fclCombineEffect, mt_evtEffect, fclShopMisc, _h_camp_menu, btlMain, game_support_late. datCalc and btlVoice experiments measured neutral/negative and were reverted, confirming the standing dead hypothesis that software double is absent from datCalc. Updated census at build/wip/W160Phantom2_census.json. LOST 0, new over-window 0, COMPILE_ERROR 0

+++ src/Main/Game/game_support_late.c
@@ -40,7 +40,7 @@ extern void* kwlnGetMainCamera(void);
 extern u32 RpRandom(void);
 extern f32 DAT_007caf38;
 extern u16 func_0016c670(s16 pcId);
-extern f32 func_00530da0(f32 value);
+extern s32 func_00530da0(f32 value);
 extern s32 sprintf(char* buffer, const char* format, ...);
 extern char* strcpy(char* destination, const char* source);
 
@@ -1505,7 +1505,7 @@ void* func_0018de60(KwlnTask* task)
                 {
                     y = 324.0f;
                 }
-                printf(D_005E4500, GS_S32(work, 8), (f64)func_00530da0(y));
+                printf(D_005E4500, GS_S32(work, 8), func_00530da0(y));
                 transition = GS_TASK(work, 0x70)->workData;
                 start.valueF[0] = GS_F32(transition, 0x38);
                 start.valueF[1] = GS_F32(transition, 0x3c);
@@ -1548,7 +1548,7 @@ void* func_0018de60(KwlnTask* task)
                 {
                     y = 324.0f;
                 }
-                printf(D_005E4500, GS_S32(work, 8), (f64)func_00530da0(y));
+                printf(D_005E4500, GS_S32(work, 8), func_00530da0(y));
                 GS_S32(work, 4) = 1;
                 transition = GS_TASK(work, 0x70)->workData;
                 start.valueF[0] = 834.0f;
